some changes based on your suggestion
I have added MLHNLabelCorrelationTest class in test packages. The
classifier have been tested by functions in MultiLabelLearnerTestBase
and passed several of them. Because the hypernetwork based approach uses
evolutionary learning strategy, the predictions of classifiers with
identical parameter setting and identical data set are not all the same.
So the classifier could not pass the testSameInputSameOutput function.
Besides, I have reviewed the code I pushed to the original repository
and translate or delete Chinese comments. Lastly, the file .project have
been removed and the name of the package been altered.