Adding implement of COCOA method
COCOA class is the implemention of COCOA method.
BinaryRelevanceUnderSampling class is the implemention of BR method with employing random undersampling strategy for each base classifier.
COCOATripleClassTransformation class is the implemention of the triple classes transformation of COCOA method.
SpreadSubsampleWithMissClassValues class is the extension of SpreadSubsample class by by adding the ability to handle instances with missing class values.
BinaryRelevanceUnderSamplingTest class and COCOATest class are the test class of BRUS and COCOA mehtod